style: promote API Logs to a standalone agent page
Mirrors the layout pattern from the crons detail and connections pages:
centered max-width shell, back-to-home link, title + primary action row,
rounded-xl entry cards. Removes the API Log settings-dialog tab and
routes the agent-home "API Logs" button through the new selection state.
